# swagger.yaml
openapi: 3.0.3
info:
  title: CareerPilot AI API
  description: REST API for the CareerPilot AI career navigation system.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    format: uuid
                  email:
                    type: string
                  accessToken:
                    type: string

  /api/auth/login:
    post:
      summary: Login user and issue session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                description: HttpOnly secure refresh token cookie
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    format: uuid
                  email:
                    type: string
                  accessToken:
                    type: string

  /api/auth/refresh:
    post:
      summary: Rotate and refresh access token using cookie
      responses:
        '200':
          description: Access token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string

  /api/profile:
    get:
      summary: Get user profile details
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved
    post:
      summary: Update user profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                college:
                  type: string
                graduationYear:
                  type: integer
                targetRoleId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Profile updated

  /api/profile/resume:
    post:
      summary: Upload resume PDF for parsing
      security:
        - BearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: Upload accepted and queue started
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  resumeId:
                    type: string
                  status:
                    type: string

  /api/careers/roles:
    get:
      summary: Get all target career roles
      responses:
        '200':
          description: List of roles with demand metrics and skill requirements

  /api/careers/assessments/{id}:
    get:
      summary: Fetch a quiz assessment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Quiz content (questions and options, answers omitted)

  /api/careers/assessments/{id}/submit:
    post:
      summary: Submit assessment quiz answers
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: array
                  items:
                    type: object
                    properties:
                      questionId:
                        type: string
                      selectedOption:
                        type: string
      responses:
        '200':
          description: Quiz graded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  score:
                    type: integer
                  passed:
                    type: boolean

  /api/roadmaps/generate:
    post:
      summary: Generate roadmap based on profile gaps and target role
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Roadmap successfully assembled and saved

  /api/roadmaps/active:
    get:
      summary: Fetch active roadmap, steps, and daily tasks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current timeline details with checklist items

  /api/tasks/{id}:
    patch:
      summary: Toggle daily task status (TODO/DONE)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [TODO, IN_PROGRESS, DONE]
      responses:
        '200':
          description: Task updated and overall score recalculated

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT